Connect user to balances and categories

+++ WebMonyAPI/WebMonyAPI/Handlers/Categories/DeleteCategoryHandler.cs
@@ -7,13 +7,20 @@ using WebMonyAPI.Entities.Categories;
 
 namespace WebMonyAPI.Handlers.Categories.Saving;
 
-public class DeleteCategoryHandler(IGenericRepository<CategoryEntity, long> repo)
+public class DeleteCategoryHandler(IGenericRepository<CategoryEntity, 
+    long> repo,
+    IIdentityService identityService)
     : IRequestHandler<DeleteCategoryCommand>
 {
     public async Task Handle(
         DeleteCategoryCommand request,
         CancellationToken cancellationToken)
     {
+        long userId = await identityService.GetUserIdAsync();
+        var entity = await repo.GetByIdAsync(request.Id);
+        if (entity == null || entity.UserId != userId)
+            throw new Exception("Category not found");
+
         await repo.DeleteAsync(request.Id);
     }
 }
